+++ src/main/java/org/example/domain/auth/token/RefreshToken.java
@@ -1,4 +1,4 @@
-package org.example.domain.auth;
+package org.example.domain.auth.token;
 
 
 import lombok.AllArgsConstructor;
